fixed tests

+++ todo_app/app.py
@@ -1,10 +1,11 @@
+from flask_login.mixins import AnonymousUserMixin
 from todo_app.data.MongoDbApi import MongoDbApi
 from todo_app.ViewModel import ViewModel
 from todo_app.data.Items import Items
 from flask import Flask, render_template, request, redirect, send_from_directory
 from todo_app.flask_config import Config
 import os
-from flask_login import LoginManager, login_required
+from flask_login import LoginManager, login_required, current_user
 from todo_app.security.Auth import Auth
 
 
@@ -17,23 +18,30 @@ def create_app(db_name=None):
     @app.route('/')
     @login_required
     def index():  # pylint:disable=unused-variable
-        item_view_model = ViewModel(
-            Items.get_items(), Items.get_current_sort_order())
+        item_view_model = None
+        if current_user.is_anonymous:
+            item_view_model = ViewModel(
+                Items.get_items(), Items.get_current_sort_order(), None)
+        else:
+            item_view_model = ViewModel(Items.get_items(
+            ), Items.get_current_sort_order(), current_user.has_write_permission())
         return render_template('index.html', view_model=item_view_model)
 
     @app.route('/', methods=['POST'])
     @login_required
     def create_item():  # pylint:disable=unused-variable
-        Items.add_item(request.form['title'])
+        if current_user.has_write_permission():
+            Items.add_item(request.form['title'])
         return redirect('/')
 
     @app.route('/actions/<action>/<id>')
     @login_required
     def perform_item_action(action, id):  # pylint:disable=unused-variable
-        if action == 'doing':
-            Items.doing_item(id)
-        elif action == 'done':
-            Items.done_item(id)
+        if current_user.has_write_permission():
+            if action == 'doing':
+                Items.doing_item(id)
+            elif action == 'done':
+                Items.done_item(id)
 
         return redirect('/')
 
@@ -46,7 +54,8 @@ def create_app(db_name=None):
     @app.route('/deleteitem/<id>')
     @login_required
     def remove_item(id):  # pylint:disable=unused-variable
-        Items.delete_item(id)
+        if current_user.has_write_permission():
+            Items.delete_item(id)
         return redirect('/')
 
     @app.route('/favicon.ico')
